pyproject.toml
[tool.poetry]
name = "play-tonight-bot"
version = "0.1.0"
description = "Discord bot for scheduling time to play games"
authors = ["Eric Chow <eric.zjp.chow@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.13.0"
discord = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
